Break conversions out of variables
