+++ variable.py
@@ -1,6 +1,7 @@
 import struct
 import sys
 from error import Error
+import converter
 
 def create(arg):
     if type(arg) is str:
@@ -141,30 +142,6 @@ MAP = {
     },
 }
 
-# Magic constants used for shifting integers to floating point numbers
-MAGIC = 32768.0
-MAGIC_AC_W_DIVISOR        = MAGIC * 800.0
-MAGIC_DC_W_DIVISOR        = MAGIC * 100.0
-MAGIC_DC_V_DIVISOR        = MAGIC * 10.0
-MAGIC_WH_MULTIPLIER       = 24.0
-MAGIC_WH_DIVISOR          = MAGIC * 100.0
-MAGIC_TEMPERATURE_DIVISOR = MAGIC
-MAGIC_PERCENT_DIVISOR     = 256.0
-
-SHUNT_NAMES = {
-    0: 'None',
-    1: 'Solar',
-    2: 'Wind',
-    3: 'Hydro',
-    4: 'Charger',
-    5: 'Load',
-    6: 'Dual',
-    7: 'Multiple SP PROs',
-    8: 'Log Only',
-    9: 'System SoC',
-    10: 'Direct SoC Input',
-}
-
 TYPES = {
     "ushort": {
         FORMAT: "<H",
@@ -183,32 +160,6 @@ TYPES = {
         WORDS: 2,
     },
 }
-
-def _convert_ac_w(raw, scales):
-    return raw * scales['CommonScaleForAcVolts'] * scales['CommonScaleForAcCurrent'] / MAGIC_AC_W_DIVISOR
-
-def _convert_ac_wh(raw, scales):
-    return raw * MAGIC_WH_MULTIPLIER * scales['CommonScaleForAcVolts'] * scales['CommonScaleForAcCurrent'] / MAGIC_WH_DIVISOR
-
-def _convert_dc_w(raw, scales):
-    return raw * scales['CommonScaleForDcVolts'] * scales['CommonScaleForDcCurrent'] / MAGIC_DC_W_DIVISOR
-
-def _convert_dc_wh(raw, scales):
-    return raw * MAGIC_WH_MULTIPLIER * scales['CommonScaleForDcVolts'] * scales['CommonScaleForDcCurrent'] / MAGIC_WH_DIVISOR
-
-def _convert_dc_v(raw, scales):
-    return raw * scales['CommonScaleForDcVolts'] / MAGIC_DC_V_DIVISOR
-
-def _convert_temperature(raw, scales):
-    return raw * scales['CommonScaleForTemperature'] / MAGIC_TEMPERATURE_DIVISOR
-
-def _convert_percent(raw, scales):
-    return raw / MAGIC_PERCENT_DIVISOR
-
-def _convert_shunt_name(raw, scales):
-    if raw in SHUNT_NAMES:
-        return SHUNT_NAMES[raw]
-    return 'Error'
 
 class Variable:
     def __init__(self, name: str, address: int, bytes: bytes=b'\x00\x00'):
@@ -251,12 +202,10 @@ class Variable:
         mem_info = MAP[self.__name]
         type_info = TYPES[self.get_type()]
         format = type_info["format"]
-        words = type_info[WORDS]
         unscaled = struct.unpack(format, self.__bytes)[0]
         if not CONVERSION in mem_info:
             return unscaled
-        scaleMethod = getattr(sys.modules[__name__], '_convert_'+mem_info[CONVERSION])
-        return scaleMethod(unscaled, scales)
+        return converter.convert(mem_info[CONVERSION], unscaled, scales)
 
     def is_known(self):
         return self.__name in MAP
